+++ tests/test_apply.py
@@ -1469,3 +1469,38 @@ def test_warning():
     """Tests if a warning is raised when lightning device binaries are not available"""
     with pytest.warns(UserWarning, match="Pre-compiled binaries for " + device_name):
         qml.device(device_name, wires=1)
+
+
+@pytest.mark.parametrize(
+    "op",
+    [
+        qml.BasisState([0, 0], wires=[0, 1]),
+        qml.QubitStateVector([0, 1, 0, 0], wires=[0, 1]),
+        qml.StatePrep([0, 1, 0, 0], wires=[0, 1]),
+    ],
+)
+@pytest.mark.parametrize("theta, phi", list(zip(THETA, PHI)))
+def test_circuit_with_stateprep(op, theta, phi, tol):
+    """Test mid-circuit StatePrep"""
+    n_qubits = 5
+    n_wires = 2
+    dev_def = qml.device("default.qubit", wires=n_qubits)
+    dev = qml.device(device_name, wires=n_qubits)
+    m = 2**n_wires
+    U = np.random.rand(m, m) + 1j * np.random.rand(m, m)
+    U, _ = np.linalg.qr(U)
+    init_state = np.random.rand(2**n_qubits) + 1j * np.random.rand(2**n_qubits)
+    init_state /= np.sqrt(np.dot(np.conj(init_state), init_state))
+
+    def circuit():
+        qml.StatePrep(init_state, wires=range(n_qubits))
+        qml.RY(theta, wires=[0])
+        qml.RY(phi, wires=[1])
+        qml.CNOT(wires=[0, 1])
+        op
+        qml.QubitUnitary(U, wires=range(2, 2 + 2 * n_wires, 2))
+        return qml.state()
+
+    circ = qml.QNode(circuit, dev)
+    circ_def = qml.QNode(circuit, dev_def)
+    assert np.allclose(circ(), circ_def(), tol)